Struct flowistry::indexed::impls::LocationIndex

pub struct LocationIndex { /* private fields */ }

Implementations

impl LocationIndex

pub const MAX_AS_U32: u32

Maximum value the index can take, as a u32.

pub const MAX: Self

Maximum value the index can take.

pub const fn from_usize(value: usize) -> Self

Creates a new index from a given usize.

Panics

Will panic if value exceeds MAX.

pub const fn from_u32(value: u32) -> Self

Creates a new index from a given u32.

Panics

Will panic if value exceeds MAX.

pub const unsafe fn from_u32_unchecked(value: u32) -> Self

Creates a new index from a given u32.

Safety

The provided value must be less than or equal to the maximum value for the newtype. Providing a value outside this range is undefined due to layout restrictions.

Prefer using from_u32.

pub const fn index(self) -> usize

Extracts the value of this index as a usize.

pub const fn as_u32(self) -> u32

Extracts the value of this index as a u32.

pub const fn as_usize(self) -> usize

Extracts the value of this index as a usize.