Struct flowistry::indexed::IndexSet

pub struct IndexSet<T: IndexedValue, S = OwnedSet<T>> { /* private fields */ }

High-level wrapper around rustc’s bitset (specifically IndexSetImpl).

This data structure ties a bitset to its IndexedDomain, enabling inputs to be represented either as values or the index of that value in the domain (see ToIndex).

To deal with references to bitsets, e.g. as returned by IndexMatrix::row_set, IndexSet has an additional type parameter for whether it contains an OwnedSet or a RefSet / MutSet.

Implementations

impl<T: IndexedValue> IndexSet<T, OwnedSet<T>>

pub fn new(domain: &Rc<T::Domain>) -> Self

impl<T, S> IndexSet<T, S>where
T: IndexedValue,
S: ToSet<T>,

pub fn to_owned(&self) -> IndexSet<T, OwnedSet<T>>

pub fn as_ref(&self) -> IndexSet<T, RefSet<'_, T>>

pub fn indices(&self) -> impl Iterator<Item = T::Index> + '_

pub fn iter(&self) -> impl Iterator<Item = &T> + '_

pub fn iter_enumerated(&self) -> impl Iterator<Item = (T::Index, &T)> + '_

pub fn contains(&self, index: impl ToIndex<T>) -> bool

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn is_superset<S2: ToSet<T>>(&self, other: &IndexSet<T, S2>) -> bool

impl<T: IndexedValue, S: ToSetMut<T>> IndexSet<T, S>

pub fn insert(&mut self, elt: impl ToIndex<T>)

pub fn union<S2: ToSet<T>>(&mut self, other: &IndexSet<T, S2>) -> bool

pub fn subtract<S2: ToSet<T>>(&mut self, other: &IndexSet<T, S2>) -> bool

pub fn intersect<S2: ToSet<T>>(&mut self, other: &IndexSet<T, S2>) -> bool

pub fn insert_all(&mut self)