Trait flowistry::indexed::IndexedDomain

pub trait IndexedDomain {
    type Value: IndexedValue;
    type Index: Idx = <Self::Value as IndexedValue>::Index;

    // Required methods
    fn value(&self, index: Self::Index) -> &Self::Value;
    fn index(&self, value: &Self::Value) -> Self::Index;
    fn contains(&self, value: &Self::Value) -> bool;
    fn as_vec(&self) -> &IndexVec<Self::Index, Self::Value>;
    fn insert(&mut self, value: Self::Value) -> Self::Index;

    // Provided methods
    fn size(&self) -> usize { ... }
    fn ensure(&mut self, value: &Self::Value) -> Self::Index { ... }
    fn as_set(
        domain: &Rc<<Self::Value as IndexedValue>::Domain>
    ) -> IndexSet<Self::Value> { ... }
}

Represents a fixed size domain of elements where each element has an index.

Enables the bidirectional mapping from element to index and vice-versa.

Required Associated Types

type Value: IndexedValue

Provided Associated Types

type Index: Idx = <Self::Value as IndexedValue>::Index

Required Methods

fn value(&self, index: Self::Index) -> &Self::Value

fn index(&self, value: &Self::Value) -> Self::Index

fn contains(&self, value: &Self::Value) -> bool

fn as_vec(&self) -> &IndexVec<Self::Index, Self::Value>

fn insert(&mut self, value: Self::Value) -> Self::Index

Provided Methods

fn size(&self) -> usize

fn ensure(&mut self, value: &Self::Value) -> Self::Index

fn as_set( domain: &Rc<<Self::Value as IndexedValue>::Domain> ) -> IndexSet<Self::Value>