Trait flowistry::indexed::ToIndex

pub trait ToIndex<T: IndexedValue> {
    fn to_index(&self, domain: &T::Domain) -> T::Index;
}

Converts an element to the index representing that element.

Useful for enabling a more flexible API where either an element or its index can be given as input. By convention, ToIndex should be implemented for an index, see to_index_impl!.

Required Methods

fn to_index(&self, domain: &T::Domain) -> T::Index