flowistry::infoflow
Function compute_flow

pub fn compute_flow<'tcx>(
    tcx: TyCtxt<'tcx>,
    body_id: BodyId,
    body_with_facts: &'tcx BodyWithBorrowckFacts<'tcx>,
) -> FlowResults<'tcx>

Computes information flow for a MIR body.

See example.rs for a complete example of how to call this function.

To get a BodyWithBorrowckFacts, you can use the get_body_with_borrowck_facts function.

See FlowResults for an explanation of how to use the return value.