flowistry::infoflow
Module mutation

Identifies the mutated places in a MIR instruction via modular approximation based on types.

Structs

ModularMutationVisitor

MIR visitor that invokes a callback for every Mutation in the visited object.

Mutation

Information about a particular mutation.

Enums

MutationStatus

Indicator of certainty about whether a place is being mutated. Used to determine whether an update should be strong or weak.

Reason

Why did this mutation occur