flowistry::mir::aliases

Struct Aliases

pub struct Aliases<'a, 'tcx> { /* private fields */ }

Data structure for computing and storing aliases.

Implementations

impl<'a, 'tcx> Aliases<'a, 'tcx>

pub fn build( tcx: TyCtxt<'tcx>, def_id: DefId, body_with_facts: &'a BodyWithBorrowckFacts<'tcx>, ) -> Self

Runs the alias analysis on a given body_with_facts.

pub fn build_with_fact_selection( tcx: TyCtxt<'tcx>, def_id: DefId, body_with_facts: &'a BodyWithBorrowckFacts<'tcx>, selector: impl Fn(RegionVid, RegionVid, <RustcFacts as FactTypes>::Point) -> bool, ) -> Self

Alternative constructor if you need to filter out certain borrowck facts.

Just use Aliases::build unless you know what you’re doing.

pub fn aliases(&self, place: Place<'tcx>) -> PlaceSet<'tcx>

Given a place, returns the set of direct places it could refer to.

For example, in the program:

let x = 1;
let y = &x;

The place *y (but NOT y) is an alias for x. Similarly, in the program:

let v = vec![1, 2, 3];
let n = &v[0];

The place *n is an alias for v (even though they have different types!).