Module flowistry::mir::utils

A potpourri of utilities for working with the MIR, primarily exposed as extension traits.

Structs

AsyncHack

PlaceCollector

SimplifyMir

MIR pass to remove instructions not important for Flowistry.

Enums

PlaceRelation

Describes the part-whole relationship between two places.

Traits

BodyExt

Extension trait for Body.

Captures

MutabilityExt

Extension trait for Mutability.

OperandExt

Extension trait for Operand.

PlaceExt

Extension trait for Place.

SpanDataExt

Extension trait for SpanData.

SpanExt

Extension trait for Span.

Functions

arg_mut_ptrs

Given the arguments to a function, returns all projections of the arguments that are mutable pointers.

arg_places

Given the arguments to a function, returns all places in the arguments.

dump_results

location_to_string

run_dot