Trait flowistry::mir::utils::PlaceExt

pub trait PlaceExt<'tcx> {
    fn make(
        local: Local,
        projection: &[PlaceElem<'tcx>],
        tcx: TyCtxt<'tcx>
    ) -> Self;
    fn from_ref(place: PlaceRef<'tcx>, tcx: TyCtxt<'tcx>) -> Self;
    fn from_local(local: Local, tcx: TyCtxt<'tcx>) -> Self;
    fn is_arg(&self, body: &Body<'tcx>) -> bool;
    fn is_direct(&self, body: &Body<'tcx>) -> bool;
    fn refs_in_projection(
        &self
    ) -> SmallVec<[(PlaceRef<'tcx>, &[PlaceElem<'tcx>]); 2]>;
    fn interior_pointers(
        &self,
        tcx: TyCtxt<'tcx>,
        body: &Body<'tcx>,
        def_id: DefId
    ) -> HashMap<RegionVid, Vec<(Place<'tcx>, Mutability)>>;
    fn interior_places(
        &self,
        tcx: TyCtxt<'tcx>,
        body: &Body<'tcx>,
        def_id: DefId
    ) -> Vec<Place<'tcx>>;
    fn interior_paths(
        &self,
        tcx: TyCtxt<'tcx>,
        body: &Body<'tcx>,
        def_id: DefId
    ) -> Vec<Place<'tcx>>;
    fn to_string(&self, tcx: TyCtxt<'tcx>, body: &Body<'tcx>) -> Option<String>;
    fn normalize(&self, tcx: TyCtxt<'tcx>, def_id: DefId) -> Place<'tcx>;
}

Extension trait for Place.

Required Methods

fn make(local: Local, projection: &[PlaceElem<'tcx>], tcx: TyCtxt<'tcx>) -> Self

Creates a new Place.

fn from_ref(place: PlaceRef<'tcx>, tcx: TyCtxt<'tcx>) -> Self

Converts a PlaceRef into an owned Place.

fn from_local(local: Local, tcx: TyCtxt<'tcx>) -> Self

Creates a new Place with an empty projection.

fn is_arg(&self, body: &Body<'tcx>) -> bool

Returns true if self is a projection of an argument local.

fn is_direct(&self, body: &Body<'tcx>) -> bool

Returns true if self could not be resolved further to another place.

This is true of places with no dereferences in the projection, or of dereferences of arguments.

fn refs_in_projection(
&self
) -> SmallVec<[(PlaceRef<'tcx>, &[PlaceElem<'tcx>]); 2]>

Returns all prefixes of self’s projection that are references, along with the suffix of the remaining projection.

fn interior_pointers(
&self,
tcx: TyCtxt<'tcx>,
body: &Body<'tcx>,
def_id: DefId
) -> HashMap<RegionVid, Vec<(Place<'tcx>, Mutability)>>

Returns all possible projections of self that are references.

The output data structure indexes the resultant places based on the region of the references.

fn interior_places(
&self,
tcx: TyCtxt<'tcx>,
body: &Body<'tcx>,
def_id: DefId
) -> Vec<Place<'tcx>>

Returns all possible projections of self that do not go through a reference, i.e. the set of fields directly in the structure referred by self.

fn interior_paths(
&self,
tcx: TyCtxt<'tcx>,
body: &Body<'tcx>,
def_id: DefId
) -> Vec<Place<'tcx>>

Returns all possible projections of self.

fn to_string(&self, tcx: TyCtxt<'tcx>, body: &Body<'tcx>) -> Option<String>

Provides a nicer debug representation of a place in terms of debug info.

fn normalize(&self, tcx: TyCtxt<'tcx>, def_id: DefId) -> Place<'tcx>

Erases/normalizes information in a place to ensure stable comparisons between places

Consider a place _1: &'1 <T as SomeTrait>::Foo[2]. We might encounter this type with a different region, e.g. &'2. We might encounter this type with a more specific type for the associated type, e.g. &'1 [i32][0]. To account for this variation, we normalize associated types, erase regions, and normalize projections.