flowistry::mir::utils

Type Alias PlaceSet

pub type PlaceSet<'tcx> = FxHashSet<Place<'tcx>>;

An unordered collections of MIR Places.

Note: this used to be implemented as an IndexSet, but in practice it was very hard to determine up-front a fixed domain of Places that was not “every possible place in the body”.

Aliased Type

pub struct PlaceSet<'tcx> { /* private fields */ }