Module flowistry::source_map

Mapping MIR instructions to source code.

Structs

GraphemeIndices

Provides efficient lookup for mapping byte indices to grapheme indices and vice-versa.

Range

Data structure for sharing spans outside rustc.

SpanTree

Interval tree data structure specialized to spans.

Spanner

Converts MIR locations to source spans using HIR information.

Enums

EnclosingHirSpans

Which parts of a HIR node’s span should be included for a matching MIR node

FunctionIdentifier

An externally-provided identifier of a function

Traits

ToSpan

Used to convert objects into a Span with access to TyCtxt

Functions

find_bodies

Finds all bodies in the current crate

find_enclosing_bodies

Finds all the bodies that enclose the given span, from innermost to outermost