Crate flowistry

This crate provides a modular information flow analysis for Rust programs, as described in the paper “Modular Information Flow Through Ownership” (Crichton et al. 2022). See example.rs for an example of how to use the Flowistry API.

Information flow describes whether one instruction or variable can affect another during a program’s execution. Information flow can be used to analyze whether secure values can leak to insecure places (“information flow control”), and to analyze which parts of a program are relevant to a given variable (“program slicing”).

This analysis uses the Rust compiler via the rustc API. Given a MIR body, the function infoflow::compute_flow computes the information flow within that body. Check out those docs for more information on the specific data structure that is computed.

If you are interested in using Flowistry, please reach out directly to Will Crichton (wcrichto@cs.stanford.edu) for questions or support. If you use Flowistry in your research, then please cite our paper:

@misc{crichton2021modular,
  title={Modular Information Flow Through Ownership},
  author={Will Crichton and Marco Patrignani and Maneesh Agrawala and Pat Hanrahan},
  year={2021},
  eprint={2111.13662},
  archivePrefix={arXiv},
  primaryClass={cs.PL}
}

Modules

extensions

Extra features for evaluating / ablating the precision of Flowistry’s algorithm.

indexed

APIs for efficiently representing values with an fixed-size domain.

infoflow

The core information flow analysis.

mir

Infrastructure for analyzing MIR that supports the information flow analysis.

source_map

test_utils

Running rustc and Flowistry in tests.

timer

Macros

block_timer

to_index_impl

Implements ToIndex for the index of an indexed type.