Module flowistry::indexed

APIs for efficiently representing values with an fixed-size domain via bit-sets.

This module is an abstraction over the rustc_index::bit_set data structures. Its main purpose is to simplify the translation between indexes and the values they represent.

Indexed values

In Flowistry, some types can only take on a specific range of values at a given time. For example, a MIR body has a fixed set of Locations. Therefore if we want to encode a set of locations (e.g. for the FlowDomain), then we can assign each Location a numeric index and use a compact bit-set instead of, say, a hash set. Concretely, this means:

Defining a type LocationOrArgIndex that implements the Idx trait.

Creating a mapping (“domain”) from Locations to LocationOrArgIndex.

Constructing a LocationSet out of a BitSet<LocationOrArgIndex>.

Abstractly, for an arbitrary type T, we perform each step with the following constructs:

Indexes are defined with the rustc_index::newtype_index macro.

The domain of T is a data structure implementing the IndexedDomain trait, such as DefaultDomain. The trait IndexedValue is implemented for T which defines an associated type T::Domain: IndexedDomain.

Then the user can construct an IndexSet<T> by providing an instance of &Rc<T::Domain>.

You can explicitly translate a T to a T::Index via IndexedDomain::index, and vice-versa with IndexedDomain::value. But this translation is often done for you. For example, most IndexSet methods use the ToIndex trait to take as input impl ToIndex<T>, and both T and T::Index implement ToIndex. So you can call IndexSet::insert with a plain T, and because IndexSet holds a pointer to its IndexedDomain, it automatically converts the T to T::Index.

Modules

impls

Structs

DefaultDomain

IndexMatrix

A mapping from R to IndexSet<C>.

IndexSet

High-level wrapper around rustc’s bitset (specifically IndexSetImpl).

MutSet

OwnedSet

RefSet

Traits

IndexSetIteratorExt

IndexedDomain

Represents a fixed size domain of elements where each element has an index.

IndexedValue

ToIndex

Converts an element to the index representing that element.

ToSet

ToSetMut

Type Aliases

IndexSetImpl

Trait Aliases

RowBounds