Module flowistry::indexed

APIs for efficiently representing values with an fixed-size domain.

Modules

impls

Structs

DefaultDomain

IndexMatrix

IndexSet

High-level wrapper around rustc’s bitset (specifically IndexSetImpl).

MutSet

OwnedSet

RefSet

Traits

IndexSetIteratorExt

IndexedDomain

Represents a fixed size domain of elements where each element has an index.

IndexedValue

ToIndex

Converts an element to the index representing that element.

ToSet

ToSetMut

Type Definitions

IndexSetImpl

Trait Aliases

RowBounds