Module flowistry::infoflow

The core information flow analysis.

The main function is compute_flow. See FlowResults and FlowDomain for an explanation of what it returns.

Modules

mutation

Identifies the mutated places in a MIR instruction via modular approximation based on types.

Structs

FlowAnalysis

Data structure that holds context for performing the information flow analysis.

Enums

Direction

Which way to look for dependencies

Functions

compute_dependencies

Computes the dependencies of a place $p$ at a location $\ell$ in a given direction.

compute_dependency_spans

Wraps compute_dependencies by translating each Location to a corresponding source Span for the location.

compute_flow

Computes information flow for a MIR body.

Type Aliases

FlowDomain

Represents the information flows at a given instruction. See FlowResults for a high-level explanation of this datatype.

FlowResults

The output of the information flow analysis.