Type Definition flowistry::infoflow::FlowDomain

pub type FlowDomain<'tcx> = IndexMatrix<Place<'tcx>, LocationOrArg>;

Represents the information flows at a given instruction. See FlowResults for a high-level explanation of this datatype.

FlowDomain represents $\Theta$ that maps from places $p$ to dependencies $\kappa$. To efficiently represent $\kappa$, a set of locations, we use the bit-set data structures in rustc_index::bit_set. However instead of using those datatypes directly, we provide several abstractions in the indexed module that have a more ergonomic interface and more efficient implementation than their bit_set counterparts.

The IndexMatrix maps from a Place to a LocationOrArgSet via the IndexMatrix::row_set method. The LocationOrArgSet is an IndexSet of locations (or arguments, see note below), which wraps a rustc_index::bit_set::HybridBitSet and has roughly the same API. The main difference is that HybridBitSet operates only on values that implement the rustc_index::vec::Idx trait (usually created via the rustc_index::newtype_index macro). The indexed module has a concept of IndexedDomain to represent the mapping from a set of values to the indexes those values — LocationOrArgDomain is the implementation for locations.

Note: arguments as dependencies

Because function arguments are never initialized, there is no “root” location for argument places. This fact poses a problem for information flow analysis: an instruction bb[0]: _2 = _1 (where _1 is an argument) would set $\Theta(\verb|_2|) = \Theta(\verb|_1|) \cup \{\verb|bb0[0]|\}$. However, $\Theta(\verb|_1|)$ would be empty, so it would be imposible to determine that _2 depends on _1. To solve this issue, we enrich the domain of locations with arguments, using the LocationOrArg type. Any dependency can be on either a location or an argument.