flowistry::mir
Module utils

A potpourri of utilities for working with the MIR, primarily exposed as extension traits.

Structs

AsyncHack

A hack to temporary hack to reduce spurious dependencies in generators arising from async functions.

Functions

arg_mut_ptrs

Given the arguments to a function, returns all projections of the arguments that are mutable pointers.

arg_places

Given the arguments to a function, returns all places in the arguments.

Type Aliases

PlaceSet

An unordered collections of MIR Places.